tikva/pyproject.toml
[project]
name = "tikva"
version = "0.1.0"
description = "Simplified robotics toolkit for CLI dataset recording"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Phospho Community"}
]
dependencies = [
    # Core utilities
    "numpy>=1.26.4",
    "scipy>=1.13.0",
    "loguru>=0.7.2",
    "pydantic>=2.10.2",
    "typer>=0.16.0",
    "rich>=14.0.0",

    # Robot control
    "dynamixel-sdk>=3.7.31",
    "feetech-servo-sdk>=1.0.0",
    "pyserial>=3.5",

    # Camera support
    "opencv-python-headless>=4.0",

    # Dataset format
    "pandas>=2.2",
    "fastparquet>=2024.11.0",
    "datasets>=3.2.0",
    "huggingface-hub>=0.28.0",

    # Configuration
    "pyyaml>=6.0",
    "toml>=0.10.2",
]

[project.optional-dependencies]
api = [
    "fastapi[standard]>=0.115.5",
    "uvicorn>=0.32.1",
    "websockets>=14.1",
]
realsense = [
    "pyrealsense2>=2.54; platform_system == 'Linux'",
    "pyrealsense2>=2.54; platform_system == 'Windows'",
    "pyrealsense2-macosx>=2.54; platform_system == 'Darwin'",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
tikva = "tikva.main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["tikva*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
